+++ Semana_3/gobarber-native/gobarberapp/react-native.config.js
@@ -1,0 +1,22 @@
+module.exports = {
+    project:{
+        ios:{
+
+        },
+        android:{
+
+        }
+    },
+    assets:[
+        './assets/fonts/'
+    ]
+}
+
+// Arquivo relacionado a rotas de assests tanto para android quando para ios
+// apos criar utilizar o comando yarn react-native link ou npx react-native link
+/**
+ * Para validar:
+ * Android - android/app/src/main/Assets/fonts
+ * IOS - ios/pastaProjeto/info.plist
+ */
+//Apos instalado rodar o comando yarn {os}