+++ Semana_3/gobarber-native/gobarberapp/src/@types/index.d.ts
@@ -1,0 +1,1 @@
+declare module '*.png'; // declara que os arquivos png sao modulos da aplicacao