+++ Semana_3/gobarber-native/gobarberapp/src/pages/SignIn/styles.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components/native";
+
+let fontRegular = "RobotoSlab-Medium";
+
+export const Container = styled.View`
+  flex: 1;
+  align-items: center;
+  /* just */
+`;
+export const Title = styled.Text`
+  font-size: 24px;
+  color: #f4ede8;
+  font-family: ${fontRegular};
+`;